Renamed all files to snake_case
Ref. #144
