+++ lib/connection/__init__.py
@@ -1,3 +1,3 @@
-from .RequestException import *
-from .Requester import *
-from .Response import *
+from .request_exception import *
+from .requester import *
+from .response import *
