+++ lib/connection/request_exception.py
