+++ lib/controller/__init__.py
@@ -1,1 +1,1 @@
-from .Controller import *
+from .controller import *
