+++ lib/core/__init__.py
@@ -1,5 +1,5 @@
-from .ArgumentParser import *
-from .Dictionary import *
-from .Fuzzer import *
-from .Path import *
-from .ReportManager import *
+from .argument_parser import *
+from .dictionary import *
+from .fuzzer import *
+from .path import *
+from .report_manager import *
