+++ lib/core/report_manager.py
