+++ lib/core/scanner.py
