+++ lib/reports/__init__.py
@@ -1,6 +1,6 @@
-from .BaseReport import *
-from .JSONReport import *
-from .PlainTextReport import *
-from .SimpleReport import *
+from .base_report import *
+from .json_report import *
+from .plain_text_report import *
+from .simple_report import *
 
 pass
