+++ lib/utils/__init__.py
@@ -1,4 +1,4 @@
-from .FileUtils import *
-from .RandomUtils import *
-from .DefaultConfigParser import *
+from .file_utils import *
+from .random_utils import *
+from .default_config_parser import *
 pass
